+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added
   - Listing webhooks
+  - Viewing details of a webhook
 
 ## [1.1.0] - 2019-10-07